move Steps enum to web directory
