+++ utils/web/types.ts
@@ -1,5 +1,12 @@
-import { Steps } from '../api/definitions/enums'
 import { FieldKey } from '../api/definitions/fields'
+
+export enum Steps {
+  STEP_1 = 'step1',
+  STEP_2 = 'step2',
+  STEP_3 = 'step3',
+  STEP_4 = 'step4',
+  STEP_5 = 'step5',
+}
 
 // Types for EligibilityPage
 export type NextClickedObject = {
